// src/models/service_order/goods/GoodOperation.ts
import type { AttachmentsOrderService } from "../../serviceOrder";
import type { Site } from "../../Site";
import type { VesselType } from "../../VesselType";
import type { Weather } from "../../Weather";

export interface GoodOperation {
    readonly id: string;
    service_order_id: string;
    vessel_type_id: string;
    vessel_type?: VesselType;
    loading_port_id?: string;
    loading_port?: Site;
    loading_facility_id?: string;
    loading_facility?: Site;
    discharge_facility_id?: string;
    discharge_facility?: Site;
    discharge_port_id?: string;
    discharge_port?: Site;
    weather_id?: string;
    weather?: Weather;
    vessel_name?: string;
    imo_number?: string;
    call_sign?: string;
    mmsi_number?: string;
    port_of_registry?: string;
    flag_state?: string;
    loa?: string;
    breadth?: string;
    depth?: string;
    gross_tonnage?: string;
    net_tonnage?: string;
    owner?: string;
    sold_to?: string;
    cargo?: string;
    description?: string;
    weight_for_transportation?: string;
    dimension?: string;
    date_of_loading?: string;
    date_of_discharge?: string;
    vessel_voyage?: string;
    flat_racks_and_position_on_board?: string;
    booking_bb?: string;
    inspector_name?: string;
    terminal_supervisor_name?: string;
    vessel_arrived?: string;
    vessel_berthed?: string;
    operations_commenced?: string;
    surveyor_at_terminal?: string;
    surveyor_on_board?: string;
    unlashing?: string;
    lifting_1?: string;
    lifting_2?: string;
    lifting_3?: string;
    lifting_4?: string;
    lifting_5?: string;
    discharge_completed?: string;
    final_inspection?: string;
    surveyor_left_terminal?: string;
    cargo_position?: string;
    lashing_equipment_and_method?: string;
    cargo_condition_on_board?: string;
    terminal_lifting_equipment?: string;
    comments_about_discharge?: string;
    final_inspection_conclusions?: string;
    temperature?: string;
    temperature_metric_unit?: string;
    relative_humidity?: string;
    sea_tide: string;
    created_at: Date;
    updated_at: Date;
    deleted_at?: Date;
    attachments?: AttachmentsOrderService[];
}